Refactor robot settings as user config
