+++ astoria/common/config/__init__.py
@@ -1,5 +1,10 @@
 """Configuration schemas for Astoria."""
 
 from .system import AstoriaConfig
+from .user import SSID_PREFIX, RobotSettings
 
-__all__ = ["AstoriaConfig"]
+__all__ = [
+    "AstoriaConfig",
+    "RobotSettings",
+    "SSID_PREFIX",
+]
